Disable ENCODE and Zebrafish support.
The requireNamespace() calls cause a WARNING if the packages are not
in the Suggests line of DESCRIPTION.  The warnign causes GitHub CI
to stop in error.  Adding the packages in Suggests causes the CI
to crash at package install time because they are only available
from http://promshift.genereg.net/CAGEr/PackageSource.  I attempted
to solve this using the following in DESCRIPTION

Remotes:
    - url::http://promshift.genereg.net/CAGEr/PackageSource/ENCODEprojectCAGE_1.0.1.tar.gz,
    - url::http://promshift.genereg.net/CAGEr/PackageSource/ZebrafishDevelopmentalCAGE_0.99.0.tar.gz

However, this causes crashes because ENCODEprojectCAGE needs
pdflatex to build.  I do not know how to support that in Win
and Mac, therefore I am giving up for the momment.

The best solution would be to get these packages in Bioconductor too.