Totally replace clusterCTSS() with distclu() and paraclu()
The expression filtering is now done by `filterLowExpCTSS` because 1) it has consequences
beyond tag cluster generation and 2) because it simplifies the argument list of the
clustering functions.

The remaining `clusterCTSS` function had a set of mutually exclusive arguments for `distclu`
and `paraclu`; I find it easier to study and use the two methods when they are taken care
of by different functions.